+++ alembic/versions/22db913c0bf7_add_items_table.py
@@ -1,0 +1,36 @@
+"""add items table
+
+Revision ID: 22db913c0bf7
+Revises:
+Create Date: 2024-09-26 15:50:38.458541
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+
+
+# revision identifiers, used by Alembic.
+revision: str = "22db913c0bf7"
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_policy(
+        table_name="items",
+        policy_name="items_permissive_all_policy_0",
+        cmd="ALL",
+        definition="PERMISSIVE",
+        expr="owner_id > NULLIF(current_setting('rls.account_id', true),'')::INTEGER",
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_policy(tablename="items", policyname="items_permissive_all_policy_0")
+    # ### end Alembic commands ###
